不具合の修正3

+++ src/app/components/FlashcardModal.tsx
@@ -3,6 +3,7 @@ import { doc, deleteDoc } from 'firebase/firestore'
 import React, { useRef, useEffect } from 'react'
 import ActionSheet from 'react-native-actionsheet'
 import { auth, db } from '../../config'
+import { Alert } from 'react-native'
 
 interface ActionSheetProps {
   visible: boolean
@@ -24,6 +25,15 @@ const FlashcardActionSheetComponent: React.FC<ActionSheetProps> = ({
   flashcardBack,
 }) => {
   const actionSheetRef = useRef<ActionSheet | null>(null)
+  
+  console.log('FlashcardActionSheetComponent', {
+    visible,
+    deckId,
+    deckName,
+    flashcardId,
+    flashcardFront,
+    flashcardBack,
+  })
 
   // const showActionSheet = () => {
   //   if (actionSheetRef.current) {
@@ -65,20 +75,35 @@ const FlashcardActionSheetComponent: React.FC<ActionSheetProps> = ({
       alert('削除に必要な情報が足りません')
       return
     }
-
-    try {
-      const flashcardRef = doc(
-        db,
-        `users/${auth.currentUser.uid}/decks/${deckId}/flashcards`,
-        flashcardId,
-      )
-      await deleteDoc(flashcardRef)
-      alert('フラッシュカードを削除しました')
-      router.push('/')
-    } catch (error) {
-      console.error('フラッシュカード削除エラー: ', error)
-      alert('削除に失敗しました')
-    }
+  
+    Alert.alert(
+      'フラッシュカードを削除しますか？',
+      '一度削除すると元に戻せません。',
+      [
+        { text: 'キャンセル', style: 'cancel' },
+        {
+          text: '削除する',
+          style: 'destructive',
+          onPress: async () => {
+            console.log('Deleting flashcard:', deckId, flashcardId)
+            try {
+              const flashcardRef = doc(
+                db,
+                `users/${auth.currentUser?.uid}/decks/${deckId}/flashcards`,
+                flashcardId,
+              )
+              await deleteDoc(flashcardRef)
+              alert('フラッシュカードを削除しました')
+              router.push('/') // 削除後にホームへ遷移
+            } catch (error) {
+              console.error('フラッシュカード削除エラー: ', error)
+              alert('削除に失敗しました')
+            }
+          },
+        },
+      ],
+      { cancelable: true }
+    )
   }
 
   useEffect(() => {
@@ -90,8 +115,8 @@ const FlashcardActionSheetComponent: React.FC<ActionSheetProps> = ({
   return (
     <ActionSheet
       ref={actionSheetRef}
-      title={'Choose an action'}
-      options={['Add', 'Edit', 'Delete', 'Cancel']}
+      title={'選択してください'}
+      options={['追加', '編集', '削除', 'キャンセル']}
       cancelButtonIndex={3}
       destructiveButtonIndex={2}
       onPress={(index) => {
